refactor: nest agent runtime authority
