+++ src/main/services/agents/integrations/runtime/event-authority.ts
@@ -1,6 +1,6 @@
 import type { AgentHookEventPayload } from "@shared/contracts/agent/session.ts";
-import type { AgentTurnStartAuthority } from "../../foreground-activity/types.ts";
-import type { AgentRuntimeSemantics } from "./types.ts";
+import type { AgentTurnStartAuthority } from "../../../foreground-activity/types.ts";
+import type { AgentRuntimeSemantics } from "../types.ts";
 
 export function resolveAgentTurnStartAuthority(
   runtime: AgentRuntimeSemantics | undefined,
