+++ tests/unit/agent-integrations/status-traces/status-trace-harness.ts
@@ -1,6 +1,6 @@
 import type { AgentStatusEvidenceDimension } from "@main/services/agents/integrations/evidence/matrix.ts";
 import { getAgentHookIntegration } from "@main/services/agents/integrations/registry.ts";
-import { resolveAgentTurnStartAuthority } from "@main/services/agents/integrations/runtime-event-authority.ts";
+import { resolveAgentTurnStartAuthority } from "@main/services/agents/integrations/runtime/event-authority.ts";
 import { createForegroundActivityAggregator } from "@main/services/foreground-activity/aggregator.ts";
 import type { AgentHookEventPayloadV3 } from "@shared/contracts/agent/session.ts";
 import { agentHookEventSchema } from "@shared/contracts/agent/session.ts";
